requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
Django==5.0.6
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-filter==24.2
django-js-asset==2.2.0
django-tenants==3.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
openpyxl==3.1.5
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
setuptools==71.0.4
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
